requirements.txt
appdirs==1.4.4
cfgv==3.1.0
cflib==0.1.11
-e git+git@github.com:Huizerd/crazyflie-suite.git@13772f5b9a8c307db3c5e610a23974702badcab1#egg=crazyflie_suite
distlib==0.3.1
filelock==3.0.12
identify==1.4.21
importlib-metadata==1.7.0
importlib-resources==3.0.0
nodeenv==1.4.0
numpy==1.19.0
pkg-resources==0.0.0
pre-commit==2.6.0
pyusb==1.0.2
PyYAML==5.3.1
six==1.15.0
toml==0.10.1
virtualenv==20.0.25
zipp==3.1.0
